Minor cleanup based on code review
